refactor: MovieList 컴포넌트 추가
